+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * vcf/lazy/record: Add info wrapper (`lazy::record::Info`).
+
 ## 0.36.0 - 2023-08-17
 
 ### Added